调试完善评论模块

+++ frontend/views/site/comments.php
@@ -150,7 +150,7 @@ stye;
 //        $where['comment_status'] = 1 ;
 
         if(empty($_COOKIE['comment_id'])){
-            var_dump($_COOKIE['comment_id']);
+//            var_dump($_COOKIE['comment_id']);
             $where = ['comment_status'=>1,'comment_post_id'=>$this->id];
         } else {
             $comment_id = $_COOKIE['comment_id'];
@@ -165,37 +165,84 @@ stye;
                 $where = ['comment_status'=>1,'comment_post_id'=>$this->id];
             }
         }
-        $comment=Comment::find()->where($where)->asArray()->all();
+        $comment=Comment::find()->where($where)->all();
 
-        var_dump($comment);
-        die();
+//        var_dump($comment);
+//        die();
 
-
+        //评价排序 按评价爹id分类
+//         $comment_group[] = "";
+        foreach($comment as $c){
+            if (!empty($c['comment_father_id'])){
+                $comment_group[$c['comment_father_id']][]=$c;
+            }
+        }
+//        echo "<br>--<br>";
+//        var_dump($comment_group);
 //todo 输出评论前，使用用户数据替换指定标签
 
         //输出评论头区部代码
         echo $this->comment_head;
-        $k=mt_rand(2,5);
+//        $k=mt_rand(2,5);
 
 //        var_dump($k,$l);
-        for($i=0;$i<$k;$i++){
+//        var_dump(array_key_exists(7,$comment_group));
+        foreach($comment as $c){
+            if(empty($c['comment_father_id'])){
+                //先输出没有父级评论的顶级评论
+//                echo '<br>comment_name='.$c['comment_name'];
 
-            echo '<li>';
-            echo $this->comment_content;
-            echo '<ul>';
-            $l=mt_rand(0,3);
-            for($j=0;$j<$l;$j++){
                 echo '<li>';
-                echo $this->comment_content;
+                $tmp = str_replace('comment_name',$c['comment_name'],$this->comment_content);
+                $tmp = str_replace('comment_content',$c['comment_content'],$tmp);
+                $tmp = str_replace('comment_time',$c['comment_time'],$tmp);
+                echo $tmp;
+                echo '<ul>';
+
+
+                if(array_key_exists($c['comment_id'],$comment_group)){
+//                  再输出该评论下的子级评论
+                    foreach($comment_group[$c['comment_id']] as $d){
+                        echo '<li>';
+                        $tmp = str_replace('comment_name',$d['comment_name'],$this->comment_content);
+                        $tmp = str_replace('comment_content',$d['comment_content'],$tmp);
+                        $tmp = str_replace('comment_time',$d['comment_time'],$tmp);
+                        echo $tmp;
+                        echo '</li>';
+
+
+                    }
+                }
+
+                echo '<li>';
+                $tmp = str_replace('comment_father_id',$c['comment_id'],$this->comment_reply);
+                echo $tmp;
+                echo '</li></ul>';
+
+
                 echo '</li>';
+
             }
-            echo '<li>';
-            echo $this->comment_reply;
-            echo '</li></ul>';
 
-
-            echo '</li>';
         }
+//        for($i=0;$i<$k;$i++){
+//
+//            echo '<li>';
+//            echo $this->comment_content;
+//            echo '<ul>';
+//            $l=mt_rand(0,3);
+//            for($j=0;$j<$l;$j++){
+//                echo '<li>';
+//                echo $this->comment_content;
+//                echo '</li>';
+//            }
+//            echo '<li>';
+//            echo $this->comment_reply;
+//            echo '</li></ul>';
+//
+//
+//            echo '</li>';
+//        }
 
         //输出评论区尾部代码
         echo $this->comment_foot;
@@ -208,10 +255,13 @@ stye;
         <div class="modal fade" id="commentsModal" tabindex="-1" role="dialog" aria-labelledby="commentsModalLabel">
             <div class="modal-dialog" role="document">
                 <div class="modal-content">
+                <div id="alert"> </div>
+
                     <div class="modal-header">
                         <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span
                                 aria-hidden="true">×</span></button>
                         <h4 class="modal-title" id="commentsModalLabel">发表评论</h4>
+
                     </div>
                     <?php $form = ActiveForm::begin(
                         ['action'=> ['comments'],]
@@ -223,7 +273,7 @@ stye;
                         <?= $form->field($model, 'comment_post_id')->hiddenInput(['value'=>$this->id])->label(false)->error(false); ?>
                         <?= $form->field($model, 'comment_name',['options'=>['class'=>'col-xs-5'],'labelOptions'=>['style'=>'float:left;padding:3px 0;width:32%;'],'inputOptions'=>['style'=>'width:68%'] ])->textInput(['maxlength' => true])->label('您的名字') ?>
                         <?= $form->field($model, 'comment_email',['options'=>['class'=>'col-xs-7'],'labelOptions'=>['style'=>'float:left;padding:3px 0;width:16%;'],'inputOptions'=>['style'=>'width:84%']])->textInput(['maxlength' => true])->label('邮箱') ?>
-                        <?= $form->field($model, 'comment_link',['options'=>['class'=>'col-xs-12'],'labelOptions'=>['style'=>'float:left;padding:3px 0;width:12%;'],'inputOptions'=>['style'=>'width:88%']])->textInput()->label('您的站点') ?>
+                        <?= $form->field($model, 'comment_link',['options'=>['class'=>'col-xs-12'],'labelOptions'=>['style'=>'float:left;padding:3px 0;width:12%;'],'inputOptions'=>['style'=>'width:88%']])->textInput()->label('您的网站') ?>
 
                         <?= $form->field($model, 'comment_content',['options'=>['class'=>'col-xs-12']])->textarea(['rows' => 3])->label('评论内容') ?>
 
@@ -232,10 +282,9 @@ stye;
                     </div>
                     <div class="modal-footer">
                         <button type="button" class="btn btn-default" data-dismiss="modal">关闭</button>
-                        //todo ajax 提交评论；前端中转保存；成功后提示；清除评论框中内容；记录用户ip；同ip免审核显示；
-                        <?= Html::submitButton('提交', ['class' => 'btn btn-primary']) ?>
-                        <button type="button" class="btn btn-primary" onclick="s();">Send message</button>
+                        <button type="button" class="btn btn-primary" onclick="s();">提交</button>
                     </div>
+
                     <?php ActiveForm::end(); ?>
                 </div>
             </div>
@@ -250,123 +299,82 @@ stye;
                     recipient = ""
                 }
                 var modal = $(this)
-                console.log(recipient)
+//                console.log(recipient)
 //                modal.find('.modal-title').text('Message To ' + recipient)
                 modal.find('.modal-body .field-comment-comment_father_id input').val(recipient)
-                console.log(modal.find('.modal-body .field-comment-comment_father_id input').get(0))
+//                console.log(modal.find('.modal-body .field-comment-comment_father_id input').get(0))
                 var str_data=$(".modal-body input").map(function(){
                     return ($(this).attr("name")+'='+$(this).val());
                 }).get().join("&") ;
-                console.log(str_data)
+//                console.log(str_data)
             })
 
-//            $.ajax({
-//
-//                var str_data=$("#dlg_form input").map(function(){
-//                    return ($(this).attr("name")+'='+$(this).val());
-//                }).get().join("&") ;
-//            type: "POST",
-//                url: "some.php",
-//                data: str_data,
-//                success: function(msg){
-//                alert( "Data Saved: " + msg );
-//            }
-//            });
         </script>
 
         <script type="text/javascript">
             //ajax post comments
-
             function s(){
+                $('#alert')[0].innerHTML= "<div class=\"alert alert-info alert-dismissable\">"
+                    +"<button type=\"button\" class=\"close\" data-dismiss=\"alert\" aria-hidden=\"true\">&times;</button>"
+                    +"<strong>评论提交中……</strong>"
+                    +"</div> ";
 
                 var d=$(".modal-body input").map(function(){
                     return ($(this).attr("name")+'='+$(this).val());
                 }).get().join("&") ;
+                d=d+"&"+$(".modal-body textarea").map(function(){
+                        return ($(this).attr("name")+'='+$(this).val());
+                    }).get().join("&") ;
 
-//                t1.innerHTML="<span class='glyphicon glyphicon-time'></span>检测中…";
-//                if(ip_t2==ip){ //如果有目标2，同步刷新
-//                    $(t2).css({"background-color":"white","color":"black"});
-//                    t2.innerHTML="<span class='glyphicon glyphicon-time'></span>检测中…";
-//                }
                 $.ajax({
                     url: "/site/comments",
                     dataType: 'json',
                     data:d,
                     type : "post",
                     success: function(html){
-                    alert(html);
-                        //  alert(t1.innerHTML);
-//                    console.log(t1.innerHTML);
-//                    console.log($(t2).css("background-color"));
-//                    console.log(key);
-                        if(html['ping_status']==1){
-                            $(t1).css({"background-color":"green","color":"white"});
-                            var out="";
-                            if("ping_time" in html){out=" "+html['ping_time']}
-                            if("ping_lost" in html){out=out+" lost("+html['ping_lost']+")"}
-                            t1.innerHTML="<span class='glyphicon glyphicon-ok'></span>"+out;
-                            if("ping_message" in html){
-                                t1.title=html['ping_message']
+                        if(html['status']==1){
+                            //提示结果
+                            $('#alert')[0].innerHTML= "<div class=\"alert alert-info alert-dismissable\">"
+                                +"<button type=\"button\" class=\"close\" data-dismiss=\"alert\" aria-hidden=\"true\">&times;</button>"
+                                +"<strong>评论提交成功！</strong>"
+                                +"</div> ";
+
+
+
+                            //清除评论内容避免重复提交
+                            if($('#comment-comment_content')[0]){
+                                $('#comment-comment_content')[0].value=""
                             }
 
 
-                            if(ip_t2==ip){
-                                t2.innerHTML=html['new_version']
-                                t2.title="当前版本："+html['new_version']+"  更新内容："+html['new_description']+"####"
-                                    +"更新前是："+html['old_version']+"  更新内容："+html['old_description']+"####"
-                                    +"缓冲区是："+html['temp_version']+"  更新内容："+html['temp_description']+"####"
-                                    +"更新时间："+html['update_time']
-                                t2.title=t2.title.replace(/\#\#\#\#/g,'\n') //将####替换成换行符，直接输出换行符会被转义成可视字符
-                                $(t2).css({"background-color":get_style(html['new_version']),"color":"white"});
-                            }
+
+                            //关闭评论窗口
+                            $(' .modal-header .close')[0].click()
+
+                            //todo 将评论内容直接添加至页面 关闭页面刷新
+                            location.reload();
+
+
                         } else {
-                            $(t1).css({"background-color":"red","color":"white"});
-                            var out="";
-                            if("ping_time" in html){out=" "+html['ping_time']}
-                            if("ping_lost" in html){out=out+" lost("+html['ping_lost']+")"}
-                            t1.innerHTML="<span class='glyphicon glyphicon-remove'></span> "+out;
-                            if("ping_message" in html){
-                                t1.title=html['ping_message']
-                            }
-                            if(ip_t2==ip){
-                                t2.innerHTML="　"
-                                $(t2).css({"background-color":"red","color":"white"});
-                            }
+                            $('#alert')[0].innerHTML= "<div class=\"alert alert-danger alert-dismissable\">"
+                            +"<button type=\"button\" class=\"close\" data-dismiss=\"alert\" aria-hidden=\"true\">&times;</button>"
+                            +"<strong>失败!</strong>    "+html['t']
+                            +"</div> ";
                         }
 
                     },
                     error: function (XMLHttpRequest, textStatus) {
-                        // alert(textStatus);
-                        t1.innerHTML="<span class='glyphicon glyphicon-exclamation-sign'></span>检测失败";
-                        if(ip_t2==ip){
-                            t2.innerHTML=t1.innerHTML
-                        }
+                        $('#alert')[0].innerHTML= "<div class=\"alert alert-danger alert-dismissable\">"
+                            +"<button type=\"button\" class=\"close\" data-dismiss=\"alert\" aria-hidden=\"true\">&times;</button>"
+                            +"<strong>sorry!评论保存失败。</strong>    "
+                            +"</div> ";
+//                        alert(XMLHttpRequest);
+//                        alert(textStatus);
                     }
                 });
 
             }
 
-
-
-            jQuery(function(){
-
-                var obj=$(".list-view");//确定数据范围
-                var list=obj.find("div[data-key]");//需要判断状态的列表
-
-                $.each(list,function(key,val){
-                    var div=$(val).find("div#ip").get(0);//需要判断状态的目标
-                    if(div!=undefined){
-                        if(div.innerHTML==1){ //判断记录状态，状态为1时才ping
-                            div.onclick=ping;
-                            div.click();
-                        } else {
-                            $(div).css({"background":"#eeeeee"})
-                            div.innerHTML="<span class='glyphicon glyphicon-ban-circle' style='color:red;'> 已禁用</span>";
-                        }
-                    }
-                })
-
-            })
         </script>
         <?php
         parent::init();
